Refactor: Relocate onboarding and setup screen components
