+++ composeApp/src/commonMain/kotlin/org/example/project/presentation/screens/setupScreens/AccountSetupEffect.kt
@@ -1,0 +1,5 @@
+package org.example.project.presentation.screens.setupScreens
+
+sealed class AccountSetupEffect {
+
+}
