+++ composeApp/src/commonMain/kotlin/org/example/project/presentation/screens/setupScreens/AccountSetupInterActionListener.kt
@@ -1,4 +1,4 @@
-package org.example.project.presentation.viewmodel.accountSetup
+package org.example.project.presentation.screens.setupScreens
 
 interface AccountSetupInterActionListener {
     fun onCategorySelected(id: Int)
