+++ composeApp/src/commonMain/kotlin/org/example/project/presentation/shared/base/BaseViewModel.kt
@@ -1,4 +1,4 @@
-package org.example.project.presentation.viewmodel.base
+package org.example.project.presentation.shared.base
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
