+++ composeApp/src/commonMain/kotlin/org/example/project/presentation/shared/base/ErrorUiState.kt
@@ -1,4 +1,4 @@
-package org.example.project.presentation.viewmodel.base
+package org.example.project.presentation.shared.base
 
 data class ErrorUiState (
     val message: String = "",
