--- composeApp/src/commonMain/kotlin/org/example/project/presentation/viewmodel/accountSetup/AccountSetupEffect.kt
@@ -1,5 +1,0 @@
-package org.example.project.presentation.viewmodel.accountSetup
-
-sealed class AccountSetupEffect {
-
-}
